+++ src/interpreter.d
@@ -274,7 +274,7 @@ class Intel8086
     }
     void Insert(ushort op, size_t offset = 0)
     {
-        uint addr = GetIPAddress + offset;
+        size_t addr = GetIPAddress + offset;
         memoryBank[addr] = op & 0xFF;
         if (op > 0xFF)
             memoryBank[++addr] = (op >> 8) & 0xFF;
